refactor(upload): extract reusable image upload into core Upload module
Move the busboy upload logic and validation helpers out of CardPlugin into
a generic API/Backend/Upload module. createUploadRouter({ subdir, ... })
saves to Missions/<mission>/<subdir>/uploads/<uuid>.<ext>, so any plugin can
mount it under its own namespace; CardPlugin now consumes it. The mount path
/api/cardplugin/upload is unchanged.

Also correct the setup.js auth comment: ensureAdmin requires an admin/lead
session and does not auto-pass under AUTH=off.
