+++ API/Backend/Upload/validate.js
@@ -1,17 +1,17 @@
-// Pure validation helpers for the CardPlugin image upload route.
+// Pure validation helpers for image upload routes (see ./uploadRouter.js).
 
-const ALLOWED_MIME_TO_EXT = {
+const IMAGE_MIME_TO_EXT = {
     'image/png': 'png',
     'image/jpeg': 'jpg',
     'image/webp': 'webp',
     'image/gif': 'gif',
 };
 
-// Map an upload mimetype to a safe file extension, or null if not allowed.
-// SVG is intentionally excluded (stored-XSS vector).
-function extensionForMime(mimeType) {
+// Map an upload mimetype to a safe file extension using the given allow-list,
+// or null if the type is not allowed.
+function extensionForMime(mimeType, allowedMimeToExt = IMAGE_MIME_TO_EXT) {
     if (typeof mimeType !== 'string') return null;
-    return ALLOWED_MIME_TO_EXT[mimeType.toLowerCase()] || null;
+    return allowedMimeToExt[mimeType.toLowerCase()] || null;
 }
 
 // Reject anything that could escape the Missions/<mission> directory.
@@ -26,4 +26,4 @@ function isValidMission(mission) {
     return true;
 }
 
-module.exports = { ALLOWED_MIME_TO_EXT, extensionForMime, isValidMission };
+module.exports = { IMAGE_MIME_TO_EXT, extensionForMime, isValidMission };
